refactor: Move more JS-specific codegen to vega-runtime.
