+++ packages/vega-parser/src/parsers/encode.js
@@ -1,10 +1,9 @@
-import adjustSpatial from './encode/adjust-spatial';
 import applyDefaults from './encode/defaults';
 import entry from './encode/entry';
 import rule from './encode/rule';
 
 import {parseExpression} from 'vega-functions';
-import {extend, isArray, stringValue} from 'vega-util';
+import {extend, isArray} from 'vega-util';
 
 export default function(encode, type, role, style, scope, params) {
   const enc = {};
@@ -20,19 +19,15 @@ export default function(encode, type, role, style, scope, params) {
 }
 
 function parseBlock(block, marktype, params, scope) {
-  const fields = {};
-  let code = 'var o=item,datum=o.datum,m=0,$;';
+  const channels = {},
+        fields = {};
 
-  for (const channel in block) {
-    const expr = expression(block[channel]);
-    code += set('o', channel, codegen(expr, scope, params, fields));
+  for (const name in block) {
+    channels[name] = codegen(expression(block[name]), scope, params, fields);
   }
 
-  code += adjustSpatial(block, marktype);
-  code += 'return m;';
-
   return {
-    $expr:   code,
+    $expr:   {marktype, channels},
     $fields: Object.keys(fields),
     $output: Object.keys(block)
   };
@@ -47,9 +42,4 @@ function codegen(code, scope, params, fields) {
   expr.$fields.forEach(name => fields[name] = 1);
   extend(params, expr.$params);
   return expr.$expr;
-}
-
-function set(obj, key, value) {
-  const o = obj + '[' + stringValue(key) + ']';
-  return `$=${value};if(${o}!==$)${o}=$,m=1;`;
 }
